requirements.txt
arcadepy>=0.1.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
claude-agent-sdk>=0.1.0
python-dotenv>=1.0.1
